requirements/xpu.txt
-r common.txt
onnxruntime>=1.23.2
intel-extension-for-pytorch
oneccl-bindings-for-pytorch
